60 clean format

+++ functional-tests/src/test/java/fake/omega/Kilo.java
@@ -25,9 +25,9 @@ package fake.omega;
  */
 
 public final class Kilo implements org.omg.CORBA.portable.IDLEntity {
-    public int tango = (int) 0;
-    public int uniform = (int) 0;
-    public int quickstep = (int) 0;
+    public int tango = 0;
+    public int uniform = 0;
+    public int quickstep = 0;
 
     public Kilo() {
     } // ctor
